Updated how long a token is good for

+++ src/main/java/us/gridiron/application/security/jwt/JwtUtils.java
@@ -23,11 +23,12 @@ public class JwtUtils {
 	private String jwtSecret;
 
 	@Value("${gridiron.app.jwtExpirationMs}")
-	private int jwtExpirationMs;
+	private Long jwtExpirationMs;
 
-	public String generateJwtToken(Authentication authentication) {
+	public String generateJwtToken(Authentication authentication)
+	{
 
-		UserDetailsImpl userPrincipal = (UserDetailsImpl) authentication.getPrincipal();
+		UserDetailsImpl userPrincipal = (UserDetailsImpl)authentication.getPrincipal();
 
 		return Jwts.builder()
 			.setSubject((userPrincipal.getUsername()))
@@ -37,26 +38,29 @@ public class JwtUtils {
 			.compact();
 	}
 
-	private Key key() {
+	private Key key()
+	{
 		return Keys.hmacShaKeyFor(Decoders.BASE64.decode(jwtSecret));
 	}
 
-	public String getUserNameFromJwtToken(String token) {
+	public String getUserNameFromJwtToken(String token)
+	{
 		return Jwts.parserBuilder().setSigningKey(key()).build()
 			.parseClaimsJws(token).getBody().getSubject();
 	}
 
-	public boolean validateJwtToken(String authToken) {
+	public boolean validateJwtToken(String authToken)
+	{
 		try {
 			Jwts.parserBuilder().setSigningKey(key()).build().parse(authToken);
 			return true;
-		} catch (MalformedJwtException e) {
+		} catch(MalformedJwtException e) {
 			logger.error("Invalid JWT token: {}", e.getMessage());
-		} catch (ExpiredJwtException e) {
+		} catch(ExpiredJwtException e) {
 			logger.error("JWT token is expired: {}", e.getMessage());
-		} catch (UnsupportedJwtException e) {
+		} catch(UnsupportedJwtException e) {
 			logger.error("JWT token is unsupported: {}", e.getMessage());
-		} catch (IllegalArgumentException e) {
+		} catch(IllegalArgumentException e) {
 			logger.error("JWT claims string is empty: {}", e.getMessage());
 		}
